Revert to Macro v1 + Help: remove keybindings and prompt_toolkit; restore simple REPL
