+++ mutants2/engine/macros.py
@@ -4,21 +4,6 @@ import re
 import time
 from pathlib import Path
 from typing import Callable, List
-from ..cli.keynames import REV_KEYPAD_CHAR
-
-
-def resolve_bound_script(store: "MacroStore", key_or_char: str) -> str | None:
-    """Return bound script for ``key_or_char`` considering keypad fallbacks."""
-    bindings = store._bindings
-    script = bindings.get(key_or_char)
-    if script:
-        return script
-    alias = REV_KEYPAD_CHAR.get(key_or_char)
-    if alias:
-        script = bindings.get(alias)
-        if script:
-            return script
-    return None
 
 
 class MacroError(Exception):
@@ -34,11 +19,6 @@ class MacroStore:
         self._macros: dict[str, str] = {}
         self.echo: bool = True
         self._call_depth: int = 0
-        self.keys_enabled: bool = True
-        self.keys_debug: bool = False
-        self.repl_mode: str = "Fallback"
-        self._bindings: dict[str, str] = {}
-        self._on_bindings_changed = None  # set by REPL
 
     # basic management -------------------------------------------------
     def add(self, name: str, script: str) -> None:
@@ -55,36 +35,13 @@ class MacroStore:
 
     def clear(self) -> None:
         self._macros.clear()
-        self._bindings.clear()
-        self._touch()
 
-    # binding management -----------------------------------------------
-    def bind(self, key: str, script: str) -> None:
-        self._bindings[key] = script
-        self._touch()
-
-    def unbind(self, key: str) -> None:
-        if key in self._bindings:
-            del self._bindings[key]
-            self._touch()
-
-    def bindings(self) -> dict[str, str]:
-        return dict(self._bindings)
-
-    def press(self, key: str, dispatch: Callable[[str], bool]) -> bool:
-        if not self.keys_enabled:
-            return False
-        script = resolve_bound_script(self, key)
-        if script is None:
-            return False
-        self.expand_and_run_script(script, dispatch)
-        return True
 
     # profile IO -------------------------------------------------------
     def save_profile(self, profile: str) -> None:
         self.MACRO_DIR.mkdir(parents=True, exist_ok=True)
         path = self.MACRO_DIR / f"{profile}.json"
-        data = {"macros": self._macros, "bindings": self._bindings, "echo": self.echo, "keys_enabled": self.keys_enabled}
+        data = {"macros": self._macros, "echo": self.echo}
         with path.open("w", encoding="utf-8") as f:
             json.dump(data, f)
 
@@ -93,23 +50,13 @@ class MacroStore:
         with path.open("r", encoding="utf-8") as f:
             data = json.load(f)
         self._macros.update(data.get("macros", {}))
-        self._bindings.update(data.get("bindings", {}))
         if "echo" in data:
             self.echo = data["echo"]
-        if "keys_enabled" in data:
-            self.keys_enabled = data["keys_enabled"]
-        self._touch()
 
     def list_profiles(self) -> List[str]:
         if not self.MACRO_DIR.is_dir():
             return []
         return sorted(p.stem for p in self.MACRO_DIR.glob("*.json"))
-
-    # internal helpers -------------------------------------------------
-    def _touch(self) -> None:
-        cb = getattr(self, "_on_bindings_changed", None)
-        if callable(cb):
-            cb()
 
     # expansion --------------------------------------------------------
     def _substitute(self, script: str, args: List[str]) -> str:
