+++ pyproject.toml
@@ -8,9 +8,7 @@ version = "0.1.0"
 description = "Simple mutants2 game"
 readme = "README.md"
 requires-python = ">=3.11"
-dependencies = [
-    "prompt_toolkit>=3.0.39",
-]
+dependencies = []
 
 [tool.pytest.ini_options]
 addopts = "-q"
